Abstracted useDynamicBottomSpacing to separate hook + fix semgrep logic in backend ci-cd
