+++ .github/workflows/backend-ci-cd.yml
@@ -53,14 +53,14 @@ jobs:
       - name: Run Semgrep for security scanning
         run: |
           python -m pip install --upgrade pip
-          pip install semgrep
-          semgrep --config p/owasp-top-ten --sarif -o ../semgrep-results.sarif
+          pip install "semgrep==1.64.0"
+          semgrep --config p/owasp-top-ten --sarif -o semgrep-results.sarif
           
       - name: Upload SARIF file
-        if: ${{ hashFiles('semgrep-results.sarif') != '' }}
+        if: ${{ hashFiles('backend/semgrep-results.sarif') != '' }}
         uses: github/codeql-action/upload-sarif@v3.25.15
         with:
-          sarif_file: semgrep-results.sarif
+          sarif_file: backend/semgrep-results.sarif
           category: semgrep
         continue-on-error: true
